+++ src/views/ShapeCreate/index.vue
@@ -53,6 +53,26 @@
               Field is required
             </p>
           </div>
+          <div
+            class="form__group"
+            :class="{'form__group--error': $v.shape.published.$error}"
+          >
+            <label>
+              <input
+                id="shape-published"
+                v-model.trim="$v.shape.published.$model"
+                placeholder="Name"
+                name="name"
+                type="checkbox"
+              >
+              Published</label>
+            <p
+              v-if="!$v.shape.definition.required"
+              class="invalid-feedback"
+            >
+              Field is required
+            </p>
+          </div>
           <div>
             <button
               class="btn btn-primary btn-rounded"
@@ -90,6 +110,7 @@ export default {
       shape: {
         name: { required },
         definition: { required },
+        published: { required },
       },
     }
   },
@@ -99,6 +120,7 @@ export default {
       shape: {
         name: null,
         definition: null,
+        published: false,
       },
       status: new Status(),
       breadcrumbs: [{